requirements.txt
streamlit==1.28.1
openai==1.3.5
python-dotenv==1.0.0
pandas==2.1.3